+++ browser/i18n/da_DK.js
@@ -225,6 +225,7 @@ module.exports = {
         "Parents": "Forældre",
         "Children": "Børn",
         "Back": "Tilbage",
+        "Fit bounds to filter": "Sæt udsnit til filter",
 
         // Base layers
         "Side-by-side mode": "Dobbelt baggrundskort",