+++ browser/i18n/en_US.js
@@ -225,6 +225,7 @@ module.exports = {
         "Parents": "Parents",
         "Children": "Children",
         "Back": "Back",
+        "Fit bounds to filter": "Fit bounds to filter",
 
         // Base layers
         "Side-by-side mode": "Side-by-side mode",